+++ src/main/java/org/example/knockin/dto/ModifyProfileLifeStyleDto.java
@@ -9,7 +9,7 @@ import java.util.List;
 public class ModifyProfileLifeStyleDto {
     @Data
     public static class Request {
-        private List<Integer> lifestyles;
+        private List<Long> lifestyles;
     }
 
     @Data
